Highlight common part
