+++ tests/onnx/quantization/test_classification_models_graph.py
@@ -17,21 +17,15 @@ import os
 
 import torch
 from torchvision import models
-import numpy as np
 import onnx
 # pylint: disable=no-member
 
 from tests.common.helpers import TEST_ROOT
-from tests.onnx.test_nncf_graph_builder import check_nx_graph
 
-from nncf.experimental.post_training.compression_builder import CompressionBuilder
-from nncf.experimental.onnx.algorithms.quantization.min_max_quantization import ONNXMinMaxQuantization
-from nncf.experimental.post_training.algorithms.quantization import MinMaxQuantizationParameters
-from nncf.experimental.post_training.algorithms.quantization import PostTrainingQuantization
-from nncf.experimental.post_training.algorithms.quantization import PostTrainingQuantizationParameters
-from nncf.experimental.post_training.api.dataloader import DataLoader
-
-from nncf.experimental.onnx.graph.nncf_graph_builder import GraphConverter
+from tests.onnx.quantization.common import min_max_quantize_model
+from tests.onnx.quantization.common import ptq_quantize_model
+from tests.onnx.quantization.common import compare_nncf_graph
+from tests.onnx.quantization.common import infer_model
 
 MODELS = [
     models.resnet18(),
@@ -60,20 +54,6 @@ INPUT_SHAPES = [
     [1, 3, 224, 224],
 ]
 
-REFERENCE_GRAPHS_TEST_ROOT = 'data/reference_graphs/quantization'
-
-
-class TestDataloader(DataLoader):
-    def __init__(self, input_shape):
-        super().__init__()
-        self.input_shape = input_shape
-
-    def __getitem__(self, item):
-        return np.squeeze(np.random.random(self.input_shape)), 0
-
-    def __len__(self):
-        return 10
-
 
 @pytest.mark.parametrize(('model', 'path_ref_graph', 'input_shape'),
                          zip(MODELS, PATH_REF_GRAPHS, INPUT_SHAPES))
@@ -87,19 +67,9 @@ def test_min_max_quantization_graph(tmp_path, model, path_ref_graph, input_shape
     torch.onnx.export(model, x, onnx_model_path, opset_version=13)
 
     original_model = onnx.load(onnx_model_path)
-
-    dataloader = TestDataloader(input_shape)
-    builder = CompressionBuilder()
-    builder.add_algorithm(ONNXMinMaxQuantization(MinMaxQuantizationParameters(number_samples=1)))
-    quantized_model = builder.apply(original_model, dataloader)
-
-    nncf_graph = GraphConverter.create_nncf_graph(quantized_model)
-    nx_graph = nncf_graph.get_graph_for_structure_analysis(extended=True)
-
-    data_dir = os.path.join(TEST_ROOT, 'onnx', REFERENCE_GRAPHS_TEST_ROOT)
-    path_to_dot = os.path.abspath(os.path.join(data_dir, path_ref_graph))
-
-    check_nx_graph(nx_graph, path_to_dot)
+    quantized_model = min_max_quantize_model(input_shape, original_model)
+    compare_nncf_graph(quantized_model, path_ref_graph)
+    infer_model(input_shape, quantized_model)
 
 
 @pytest.mark.parametrize(('model', 'path_ref_graph', 'input_shape'),
@@ -114,16 +84,6 @@ def test_post_training_quantization_graph(tmp_path, model, path_ref_graph, input
     torch.onnx.export(model, x, onnx_model_path, opset_version=13)
 
     original_model = onnx.load(onnx_model_path)
-
-    dataloader = TestDataloader(input_shape)
-    builder = CompressionBuilder()
-    builder.add_algorithm(PostTrainingQuantization(PostTrainingQuantizationParameters(number_samples=1)))
-    quantized_model = builder.apply(original_model, dataloader)
-
-    nncf_graph = GraphConverter.create_nncf_graph(quantized_model)
-    nx_graph = nncf_graph.get_graph_for_structure_analysis(extended=True)
-
-    data_dir = os.path.join(TEST_ROOT, 'onnx', REFERENCE_GRAPHS_TEST_ROOT)
-    path_to_dot = os.path.abspath(os.path.join(data_dir, path_ref_graph))
-
-    check_nx_graph(nx_graph, path_to_dot)
+    quantized_model = ptq_quantize_model(input_shape, original_model)
+    compare_nncf_graph(quantized_model, path_ref_graph)
+    infer_model(input_shape, quantized_model)
