+++ tests/onnx/quantization/test_detection_models_graph.py
@@ -14,21 +14,16 @@
 import pytest
 
 import os
-import tempfile
+
 import requests
 
 import onnx
 
 from tests.common.helpers import TEST_ROOT
-from tests.onnx.test_nncf_graph_builder import check_nx_graph
 
-from tests.onnx.quantization.test_classification_models_graph import TestDataloader
-
-from nncf.experimental.post_training.compression_builder import CompressionBuilder
-from nncf.experimental.onnx.algorithms.quantization.min_max_quantization import ONNXMinMaxQuantization
-from nncf.experimental.post_training.algorithms.quantization import MinMaxQuantizationParameters
-
-from nncf.experimental.onnx.graph.nncf_graph_builder import GraphConverter
+from tests.onnx.quantization.common import min_max_quantize_model
+from tests.onnx.quantization.common import compare_nncf_graph
+from tests.onnx.quantization.common import infer_model
 
 MODELS_NAME = [
     'yolov2-coco-9',
@@ -50,8 +45,6 @@ INPUT_SHAPES = [
     [1, 3, 416, 416],
 ]
 
-REFERENCE_GRAPHS_TEST_ROOT = 'data/reference_graphs/quantization'
-
 
 @pytest.mark.parametrize(('model_name', 'model_url', 'path_ref_graph', 'input_shape'),
                          zip(MODELS_NAME, MODELS_URL, PATH_REF_GRAPHS, INPUT_SHAPES))
@@ -66,24 +59,6 @@ def test_min_max_quantization_graph(tmp_path, model_name, model_url, path_ref_gr
         f.write(r.content)
 
     original_model = onnx.load(onnx_model_path)
-
-    dataloader = TestDataloader(input_shape)
-    builder = CompressionBuilder()
-    builder.add_algorithm(ONNXMinMaxQuantization(MinMaxQuantizationParameters(number_samples=1)))
-    quantized_model = builder.apply(original_model, dataloader)
-
-    nncf_graph = GraphConverter.create_nncf_graph(quantized_model)
-    nx_graph = nncf_graph.get_graph_for_structure_analysis(extended=True)
-
-    data_dir = os.path.join(TEST_ROOT, 'onnx', REFERENCE_GRAPHS_TEST_ROOT)
-    path_to_dot = os.path.abspath(os.path.join(data_dir, path_ref_graph))
-
-    check_nx_graph(nx_graph, path_to_dot)
-
-    with tempfile.NamedTemporaryFile() as temporary_model:
-        onnx.save(quantized_model, temporary_model.name)
-
-        sess = rt.InferenceSession(temporary_model.name, providers=['OpenVINOExecutionProvider'])
-        _input = np.random.random(input_shape)
-        input_name = sess.get_inputs()[0].name
-        _ = sess.run([], {input_name: _input.astype(np.float32)})
+    quantized_model = min_max_quantize_model(input_shape, original_model)
+    compare_nncf_graph(quantized_model, path_ref_graph)
+    infer_model(input_shape, quantized_model)
